+++ src/include/program_util.hpp
@@ -31,6 +31,8 @@ class ProgramUtil {
 
   static bool isCommutative(Operation::Type t);
 
+  static bool isAdditive(Operation::Type t);
+
   static bool hasIndirectOperand(const Operation &op);
 
   static bool areIndependent(const Operation &op1, const Operation &op2);